requirements.txt
streamlit
duckdb
pandas
pillow
